+++ apps/themovie/src/components/MovieCard/renderMoviesCard.js
@@ -1,0 +1,7 @@
+import MovieCard from './MovieCard'
+
+
+
+const renderMoviesCard = ( movie, index ) => ( <MovieCard key={`movie-${index}`} {...movie} /> )
+
+export default renderMoviesCard
